Declare whoWon and update playerSelection
The whoWon function was declared but is not yet fully working. The idea
is beginning to take shape.
Functionality was added to the playerSelection function to display
'Invalid choice' if 'rock', 'paper', or 'scissors' was not chosen. I do
know that a loop would solve this problem better, however I have not
gotten there yet in the curriculum and want to focus on solutions
within my current understanding.