Add a button which starts the game
This is the final result I have been wanting. It feels so weird to just
get bombarded with prompts out of nowhere. This makes it feel a little
more like an actual game.
There may be a couple more optimizations I can add from here, but
overall I am very pleased with the game as it stands.