Fix bug that allowed for invalid answers
Previously the code assumed that the user would type in 'rock',
'paper', or 'scissors' in some combination of capitalized spelling.
Thus it was set up so that the default value for playerChoice would be
'Scissors' if playerChoice was not previously 'rock' or 'paper'. This
meant that anything other than rock or paper would automatically be
scissors. kjhkjh would become scissors. This obviously is not correct.
Additionally, some fluff code was discovered in the same area and
removed.