Add console.logs to improve viewability
Sometimes the prompt window would be in the way of viewing the console.
Now the first time a page is loaded you will be able to see everything
no matter the proportions (hopefully). This also has a secondary
benefit of delineating between games in the console by provide space
between them.